LayoutTests/fast/canvas/webgl/bad-arguments-test.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
<script src="resources/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
if (window.internals)
    window.internals.settings.setWebGLErrorsToConsoleEnabled(false);

var wtu = WebGLTestUtils;
description("Tests calling WebGL APIs with wrong argument types");

var context = wtu.create3DContext();
var program = wtu.loadStandardProgram(context);
var shader = wtu.loadStandardVertexShader(context);
var shouldGenerateGLError = wtu.shouldGenerateGLError;

assertMsg(program != null, "Program Compiled");
assertMsg(shader != null, "Shader Compiled");

var loc = context.getUniformLocation(program, "u_modelViewProjMatrix");
assertMsg(loc != null, "getUniformLocation succeeded");

var arguments = [
  { value: "foo",
    throw: true },
  { value: 0,
    throw: true },
  { value: null,
    throw: false },
  { value: undefined,
    throw: false }
];

var argument;

function shouldBeEmptyString(command) {
  shouldBe(command, "''");
}

for (var i = 0; i < arguments.length; ++i) {
  var func, func2, func3;
  if (arguments[i].throw) {
    func = shouldThrow;
    func2 = shouldThrow;
    func3 = shouldThrow;
  } else {
    func = shouldBeUndefined;
    func2 = shouldBeNull;
    func3 = shouldBeEmptyString;
  }
  argument = arguments[i].value;
  func("context.compileShader(argument)");
  func("context.linkProgram(argument)");
  func("context.attachShader(program, argument)");
  func("context.attachShader(argument, shader)");
  func("context.detachShader(program, argument)");
  func("context.detachShader(argument, shader)");
  func("context.useProgram(argument)");
  func("context.shaderSource(argument, 'foo')");
  func("context.bindAttribLocation(argument, 0, 'foo')");
  func("context.bindBuffer(context.ARRAY_BUFFER, argument)");
  func("context.bindFramebuffer(context.FRAMEBUFFER, argument)");
  func("context.bindRenderbuffer(context.RENDERBUFFER, argument)");
  func("context.bindTexture(context.TEXTURE_2D, argument)");
  func("context.framebufferRenderbuffer(context.FRAMEBUFFER, context.DEPTH_ATTACHMENT, context.RENDERBUFFER, argument)");
  func("context.framebufferTexture2D(context.FRAMEBUFFER, context.COLOR_ATTACHMENT0, context.TEXTURE_2D, argument, 0)");
  func("context.uniform2fv(argument, new Float32Array([0.0, 0.0]))");
  func("context.uniform2iv(argument, new Int32Array([0, 0]))");
  func("context.uniformMatrix2fv(argument, false, new Float32Array([0.0, 0.0, 0.0, 0.0]))");

  func2("context.getProgramParameter(argument, 0)");
  func2("context.getShaderParameter(argument, 0)");
  func2("context.getUniform(argument, loc)");
  func2("context.getUniform(program, argument)");
  func2("context.getUniformLocation(argument, 'u_modelViewProjMatrix')");

  func3("context.getProgramInfoLog(argument)");
  func3("context.getShaderInfoLog(argument)");
  func3("context.getShaderSource(argument)");
}
</script>

<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
